+++ client/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.4] - 2026-06-03
+
+### Added
+
+- `Client::handler() -> &impl ClientHandler`
+
 ## [0.9.3] - 2026-05-26
 
 ### Added